品类阶梯价修复 服务费系数为小数

+++ src/main/java/com/wanhutong/backend/modules/biz/service/shelf/BizVarietyFactorService.java
@@ -53,7 +53,7 @@ public class BizVarietyFactorService extends CrudService<BizVarietyFactorDao, Bi
 				}
 				varietyFactor.setMinQty(Integer.parseInt(minQtyArr[i]));
 				varietyFactor.setMaxQty(Integer.parseInt(maxQtyArr[i]));
-				varietyFactor.setServiceFactor(Integer.parseInt(serviceFactorArr[i]));
+				varietyFactor.setServiceFactor(Double.parseDouble(serviceFactorArr[i]));
 				super.save(varietyFactor);
 			}
 		}
